move extensions adders to the extension package
