+++ java/src/main/java/com/powsybl/dataframe/network/extensions/EntsoeAreaDataframeAdder.java
@@ -4,25 +4,26 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dataframe.network.adders;
+package com.powsybl.dataframe.network.extensions;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dataframe.SeriesMetadata;
+import com.powsybl.dataframe.network.adders.AbstractSimpleAdder;
+import com.powsybl.dataframe.network.adders.SeriesUtils;
 import com.powsybl.dataframe.update.StringSeries;
 import com.powsybl.dataframe.update.UpdatingDataframe;
-import com.powsybl.entsoe.util.XnodeAdder;
-import com.powsybl.iidm.network.DanglingLine;
+import com.powsybl.entsoe.util.EntsoeAreaAdder;
+import com.powsybl.entsoe.util.EntsoeGeographicalCode;
 import com.powsybl.iidm.network.Network;
+import com.powsybl.iidm.network.Substation;
 
 import java.util.Collections;
 import java.util.List;
 
-import static com.powsybl.dataframe.network.adders.SeriesUtils.applyIfPresent;
-
 /**
  * @author Christian Biasuzzi <christian.biasuzzi@soft.it>
  */
-public class XnodeDataframeAdder extends AbstractSimpleAdder {
+public class EntsoeAreaDataframeAdder extends AbstractSimpleAdder {
 
     private static final List<SeriesMetadata> METADATA = List.of(
             SeriesMetadata.stringIndex("id"),
@@ -34,31 +35,31 @@ public class XnodeDataframeAdder extends AbstractSimpleAdder {
         return Collections.singletonList(METADATA);
     }
 
-    private static class XnodeSeries {
+    private static class EntsoeAreaSeries {
 
         private final StringSeries id;
         private final StringSeries code;
 
-        XnodeSeries(UpdatingDataframe dataframe) {
+        EntsoeAreaSeries(UpdatingDataframe dataframe) {
             this.id = dataframe.getStrings("id");
             this.code = dataframe.getStrings("code");
         }
 
         void create(Network network, int row) {
             String id = this.id.get(row);
-            DanglingLine l = network.getDanglingLine(id);
-            if (l == null) {
-                throw new PowsyblException("Invalid dangling line id : could not find " + id);
+            Substation s = network.getSubstation(id);
+            if (s == null) {
+                throw new PowsyblException("Invalid substation id : could not find " + id);
             }
-            var adder = l.newExtension(XnodeAdder.class);
-            applyIfPresent(code, row, adder::withCode);
+            var adder = s.newExtension(EntsoeAreaAdder.class);
+            SeriesUtils.applyIfPresent(code, row, c -> adder.withCode(EntsoeGeographicalCode.valueOf(c)));
             adder.add();
         }
     }
 
     @Override
     public void addElements(Network network, UpdatingDataframe dataframe) {
-        XnodeSeries series = new XnodeSeries(dataframe);
+        EntsoeAreaSeries series = new EntsoeAreaSeries(dataframe);
         for (int row = 0; row < dataframe.getRowCount(); row++) {
             series.create(network, row);
         }
