+++ java/src/main/java/com/powsybl/dataframe/network/extensions/HvdcAngleDroopActivePowerControlDataframeProvider.java
@@ -10,7 +10,6 @@ import com.google.auto.service.AutoService;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dataframe.network.NetworkDataframeMapper;
 import com.powsybl.dataframe.network.NetworkDataframeMapperBuilder;
-import com.powsybl.dataframe.network.adders.HvdcAngleDroopActivePowerControlDataframeAdder;
 import com.powsybl.dataframe.network.adders.NetworkElementAdder;
 import com.powsybl.iidm.network.HvdcLine;
 import com.powsybl.iidm.network.Network;
