+++ java/src/main/java/com/powsybl/dataframe/network/extensions/HvdcOperatorActivePowerRangeDataframeAdder.java
@@ -4,34 +4,31 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dataframe.network.adders;
+package com.powsybl.dataframe.network.extensions;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dataframe.SeriesMetadata;
+import com.powsybl.dataframe.network.adders.AbstractSimpleAdder;
+import com.powsybl.dataframe.network.adders.SeriesUtils;
 import com.powsybl.dataframe.update.DoubleSeries;
-import com.powsybl.dataframe.update.IntSeries;
 import com.powsybl.dataframe.update.StringSeries;
 import com.powsybl.dataframe.update.UpdatingDataframe;
 import com.powsybl.iidm.network.HvdcLine;
 import com.powsybl.iidm.network.Network;
-import com.powsybl.iidm.network.extensions.HvdcAngleDroopActivePowerControlAdder;
+import com.powsybl.iidm.network.extensions.HvdcOperatorActivePowerRangeAdder;
 
 import java.util.Collections;
 import java.util.List;
 
-import static com.powsybl.dataframe.network.adders.SeriesUtils.applyBooleanIfPresent;
-import static com.powsybl.dataframe.network.adders.SeriesUtils.applyIfPresent;
-
 /**
  * @author Christian Biasuzzi <christian.biasuzzi@soft.it>
  */
-public class HvdcAngleDroopActivePowerControlDataframeAdder extends AbstractSimpleAdder {
+public class HvdcOperatorActivePowerRangeDataframeAdder extends AbstractSimpleAdder {
 
     private static final List<SeriesMetadata> METADATA = List.of(
             SeriesMetadata.stringIndex("id"),
-            SeriesMetadata.doubles("droop"),
-            SeriesMetadata.doubles("p0"),
-            SeriesMetadata.booleans("enabled")
+            SeriesMetadata.doubles("opr_from_cs1_to_cs2"),
+            SeriesMetadata.doubles("opr_from_cs2_to_cs1")
             );
 
     @Override
@@ -39,18 +36,16 @@ public class HvdcAngleDroopActivePowerControlDataframeAdder extends AbstractSimp
         return Collections.singletonList(METADATA);
     }
 
-    private static class HvdcAngleDroopActivePowerControlSeries {
+    private static class HvdcOperatorActivePowerRangeSeries {
 
         private final StringSeries id;
-        private final DoubleSeries droop;
-        private final DoubleSeries p0;
-        private final IntSeries enabled;
+        private final DoubleSeries oprFromCS1toCS2;
+        private final DoubleSeries oprFromCS2toCS1;
 
-        HvdcAngleDroopActivePowerControlSeries(UpdatingDataframe dataframe) {
+        HvdcOperatorActivePowerRangeSeries(UpdatingDataframe dataframe) {
             this.id = dataframe.getStrings("id");
-            this.droop = dataframe.getDoubles("droop");
-            this.p0 = dataframe.getDoubles("p0");
-            this.enabled = dataframe.getInts("enabled");
+            this.oprFromCS1toCS2 = dataframe.getDoubles("opr_from_cs1_to_cs2");
+            this.oprFromCS2toCS1 = dataframe.getDoubles("opr_from_cs2_to_cs1");
         }
 
         void create(Network network, int row) {
@@ -59,17 +54,16 @@ public class HvdcAngleDroopActivePowerControlDataframeAdder extends AbstractSimp
             if (l == null) {
                 throw new PowsyblException("Invalid hvdc line id : could not find " + id);
             }
-            var adder = l.newExtension(HvdcAngleDroopActivePowerControlAdder.class);
-            applyIfPresent(droop, row, x -> adder.withDroop((float) x));
-            applyIfPresent(p0, row, x -> adder.withP0((float) x));
-            applyBooleanIfPresent(enabled, row, adder::withEnabled);
+            var adder = l.newExtension(HvdcOperatorActivePowerRangeAdder.class);
+            SeriesUtils.applyIfPresent(oprFromCS1toCS2, row, x -> adder.withOprFromCS1toCS2((float) x));
+            SeriesUtils.applyIfPresent(oprFromCS2toCS1, row, x -> adder.withOprFromCS2toCS1((float) x));
             adder.add();
         }
     }
 
     @Override
     public void addElements(Network network, UpdatingDataframe dataframe) {
-        HvdcAngleDroopActivePowerControlSeries series = new HvdcAngleDroopActivePowerControlSeries(dataframe);
+        HvdcOperatorActivePowerRangeSeries series = new HvdcOperatorActivePowerRangeSeries(dataframe);
         for (int row = 0; row < dataframe.getRowCount(); row++) {
             series.create(network, row);
         }
